src/features/social/hooks/useSocialActions.ts
import { useCallback } from 'react';
import { useStore } from '../../../store';
import { usePoints } from '../../../hooks/usePoints';

export const useSocialActions = () => {
  const { user } = useStore();
  const { awardPoints } = usePoints();

  const connectWithUser = useCallback((userId: string) => {
    if (!user) return;
    
    // TODO: Implement actual connection logic
    awardPoints(10); // Award points for networking
  }, [user, awardPoints]);

  const shareEvent = useCallback((eventId: string, platform: 'twitter' | 'facebook' | 'linkedin') => {
    if (!user) return;
    
    // TODO: Implement actual sharing logic
    awardPoints(15); // Award points for sharing
  }, [user, awardPoints]);

  return {
    connectWithUser,
    shareEvent,
  };
};